+++ backend/modules/brain/dto/inputs.py
@@ -1,3 +1,4 @@
+from enum import Enum
 from typing import Optional
 from uuid import UUID
 
@@ -14,6 +15,11 @@ from pydantic import BaseModel, Extra
 logger = get_logger(__name__)
 
 
+class IntegrationType(str, Enum):
+    CUSTOM = "custom"
+    SYNC = "sync"
+
+
 class CreateApiBrainDefinition(BaseModel, extra=Extra.forbid):
     method: ApiBrainAllowedMethods
     url: str
@@ -28,6 +34,9 @@ class CreateIntegrationBrain(BaseModel, extra=Extra.forbid):
     integration_name: str
     integration_logo_url: str
     connection_settings: dict
+    integration_type: IntegrationType
+    description: str
+    max_files: int
 
 
 class BrainIntegrationSettings(BaseModel, extra=Extra.forbid):
@@ -43,12 +52,12 @@ class CreateBrainProperties(BaseModel, extra=Extra.forbid):
     name: Optional[str] = "Default brain"
     description: str = "This is a description"
     status: Optional[str] = "private"
-    model: Optional[str]
+    model: Optional[str] = None
     temperature: Optional[float] = 0.0
     max_tokens: Optional[int] = 2000
     prompt_id: Optional[UUID] = None
     brain_type: Optional[BrainType] = BrainType.DOC
-    brain_definition: Optional[CreateApiBrainDefinition]
+    brain_definition: Optional[CreateApiBrainDefinition] = None
     brain_secrets_values: Optional[dict] = {}
     connected_brains_ids: Optional[list[UUID]] = []
     integration: Optional[BrainIntegrationSettings] = None
@@ -61,14 +70,14 @@ class CreateBrainProperties(BaseModel, extra=Extra.forbid):
 
 
 class BrainUpdatableProperties(BaseModel):
-    name: Optional[str]
-    description: Optional[str]
-    temperature: Optional[float]
-    model: Optional[str]
-    max_tokens: Optional[int]
-    status: Optional[str]
-    prompt_id: Optional[UUID]
-    brain_definition: Optional[ApiBrainDefinitionEntity]
+    name: Optional[str] = None
+    description: Optional[str] = None
+    temperature: Optional[float] = None
+    model: Optional[str] = None
+    max_tokens: Optional[int] = None
+    status: Optional[str] = None
+    prompt_id: Optional[UUID] = None
+    brain_definition: Optional[ApiBrainDefinitionEntity] = None
     connected_brains_ids: Optional[list[UUID]] = []
     integration: Optional[BrainIntegrationUpdateSettings] = None