+++ backend/modules/brain/knowledge_brain_qa.py
@@ -1,24 +1,28 @@
-import asyncio
 import json
-from typing import AsyncIterable, Awaitable, List, Optional
+from typing import AsyncIterable, List, Optional
 from uuid import UUID
 
 from langchain.callbacks.streaming_aiter import AsyncIteratorCallbackHandler
-from langchain.chains import ConversationalRetrievalChain
 from llm.utils.format_chat_history import format_chat_history
 from llm.utils.get_prompt_to_use import get_prompt_to_use
 from llm.utils.get_prompt_to_use_id import get_prompt_to_use_id
 from logger import get_logger
 from models import BrainSettings
+from models.user_usage import UserUsage
 from modules.brain.qa_interface import QAInterface
 from modules.brain.rags.quivr_rag import QuivrRAG
 from modules.brain.rags.rag_interface import RAGInterface
 from modules.brain.service.brain_service import BrainService
+from modules.chat.controller.chat.utils import (
+    find_model_and_generate_metadata,
+    update_user_usage,
+)
 from modules.chat.dto.chats import ChatQuestion, Sources
 from modules.chat.dto.inputs import CreateChatHistory
 from modules.chat.dto.outputs import GetChatHistoryOutput
 from modules.chat.service.chat_service import ChatService
-from pydantic import BaseModel
+from pydantic import BaseModel, ConfigDict
+from pydantic_settings import BaseSettings
 from repository.files.generate_file_signed_url import generate_file_signed_url
 
 logger = get_logger(__name__)
@@ -38,7 +42,7 @@ def is_valid_uuid(uuid_to_test, version=4):
     return str(uuid_obj) == uuid_to_test
 
 
-def generate_source(result, brain):
+def generate_source(source_documents, brain_id):
     # Initialize an empty list for sources
     sources_list: List[Sources] = []
 
@@ -46,13 +50,13 @@ def generate_source(result, brain):
     generated_urls = {}
 
     # Get source documents from the result, default to an empty list if not found
-    source_documents = result.get("source_documents", [])
 
     # If source documents exist
     if source_documents:
         logger.info(f"Source documents found: {source_documents}")
         # Iterate over each document
         for doc in source_documents:
+            logger.info("Document: %s", doc)
             # Check if 'url' is in the document metadata
             logger.info(f"Metadata 1: {doc.metadata}")
             is_url = (
@@ -76,7 +80,7 @@ def generate_source(result, brain):
             if is_url:
                 source_url = doc.metadata["original_file_name"]
             else:
-                file_path = f"{brain.brain_id}/{doc.metadata['file_name']}"
+                file_path = f"{brain_id}/{doc.metadata['file_name']}"
                 # Check if the URL has already been generated
                 if file_path in generated_urls:
                     source_url = generated_urls[file_path]
@@ -111,40 +115,44 @@ class KnowledgeBrainQA(BaseModel, QAInterface):
     Each have the same prompt template, which is defined in the `prompt_template` property.
     """
 
-    class Config:
-        """Configuration of the Pydantic Object"""
-
-        arbitrary_types_allowed = True
+    model_config = ConfigDict(arbitrary_types_allowed=True)
 
     # Instantiate settings
-    brain_settings = BrainSettings()  # type: ignore other parameters are optional
+    brain_settings: BaseSettings = BrainSettings()
 
     # Default class attributes
     model: str = None  # pyright: ignore reportPrivateUsage=none
     temperature: float = 0.1
     chat_id: str = None  # pyright: ignore reportPrivateUsage=none
-    brain_id: str  # pyright: ignore reportPrivateUsage=none
+    brain_id: str = None  # pyright: ignore reportPrivateUsage=none
     max_tokens: int = 2000
     max_input: int = 2000
     streaming: bool = False
-    knowledge_qa: Optional[RAGInterface]
+    knowledge_qa: Optional[RAGInterface] = None
+    user_id: str = None
+    user_email: str = None
+    user_usage: Optional[UserUsage] = None
+    user_settings: Optional[dict] = None
+    models_settings: Optional[List[dict]] = None
     metadata: Optional[dict] = None
 
     callbacks: List[AsyncIteratorCallbackHandler] = (
         None  # pyright: ignore reportPrivateUsage=none
     )
 
-    prompt_id: Optional[UUID]
+    prompt_id: Optional[UUID] = None
 
     def __init__(
         self,
         model: str,
         brain_id: str,
         chat_id: str,
-        max_tokens: int,
         streaming: bool = False,
         prompt_id: Optional[UUID] = None,
         metadata: Optional[dict] = None,
+        user_id: str = None,
+        user_email: str = None,
+        cost: int = 100,
         **kwargs,
     ):
         super().__init__(
@@ -162,8 +170,17 @@ class KnowledgeBrainQA(BaseModel, QAInterface):
             streaming=streaming,
             **kwargs,
         )
-        self.metadata = metadata
-        self.max_tokens = max_tokens
+        self.user_id = user_id
+        self.user_email = user_email
+        self.user_usage = UserUsage(
+            id=user_id,
+            email=user_email,
+        )
+        self.user_settings = self.user_usage.get_user_settings()
+
+        # Get Model settings for the user
+        self.models_settings = self.user_usage.get_model_settings()
+        self.increase_usage_user()
 
     @property
     def prompt_to_use(self):
@@ -180,38 +197,57 @@ class KnowledgeBrainQA(BaseModel, QAInterface):
         else:
             return None
 
+    def increase_usage_user(self):
+        # Raises an error if the user has consumed all of of his credits
+
+        update_user_usage(
+            usage=self.user_usage,
+            user_settings=self.user_settings,
+            cost=self.calculate_pricing(),
+        )
+
+    def calculate_pricing(self):
+
+        logger.info("Calculating pricing")
+        logger.info(f"Model: {self.model}")
+        logger.info(f"User settings: {self.user_settings}")
+        logger.info(f"Models settings: {self.models_settings}")
+        model_to_use = find_model_and_generate_metadata(
+            self.chat_id,
+            self.model,
+            self.user_settings,
+            self.models_settings,
+        )
+
+        self.model = model_to_use.name
+        self.max_input = model_to_use.max_input
+        self.max_tokens = model_to_use.max_output
+        user_choosen_model_price = 1000
+
+        for model_setting in self.models_settings:
+            if model_setting["name"] == self.model:
+                user_choosen_model_price = model_setting["price"]
+
+        return user_choosen_model_price
+
     def generate_answer(
         self, chat_id: UUID, question: ChatQuestion, save_answer: bool = True
     ) -> GetChatHistoryOutput:
-        transformed_history = format_chat_history(
-            chat_service.get_chat_history(self.chat_id)
+        conversational_qa_chain = self.knowledge_qa.get_chain()
+        transformed_history, streamed_chat_history = (
+            self.initialize_streamed_chat_history(chat_id, question)
         )
-
-        # The Chain that combines the question and answer
-        qa = ConversationalRetrievalChain(
-            retriever=self.knowledge_qa.get_retriever(),
-            combine_docs_chain=self.knowledge_qa.get_doc_chain(
-                streaming=False,
-            ),
-            question_generator=self.knowledge_qa.get_question_generation_llm(),
-            verbose=False,
-            rephrase_question=False,
-            return_source_documents=True,
-        )
-
-        prompt_content = (
-            self.prompt_to_use.content if self.prompt_to_use else QUIVR_DEFAULT_PROMPT
-        )
-
-        model_response = qa(
+        model_response = conversational_qa_chain.invoke(
             {
                 "question": question.question,
                 "chat_history": transformed_history,
-                "custom_personality": prompt_content,
+                "custom_personality": (
+                    self.prompt_to_use.content if self.prompt_to_use else None
+                ),
             }
         )
 
-        answer = model_response["answer"]
+        answer = model_response["answer"].content
 
         brain = brain_service.get_brain_by_id(self.brain_id)
 
@@ -262,124 +298,77 @@ class KnowledgeBrainQA(BaseModel, QAInterface):
     async def generate_stream(
         self, chat_id: UUID, question: ChatQuestion, save_answer: bool = True
     ) -> AsyncIterable:
-        history = chat_service.get_chat_history(self.chat_id)
-        callback = AsyncIteratorCallbackHandler()
-        self.callbacks = [callback]
-
-        # The Chain that combines the question and answer
-        qa = ConversationalRetrievalChain(
-            retriever=self.knowledge_qa.get_retriever(),
-            combine_docs_chain=self.knowledge_qa.get_doc_chain(
-                callbacks=self.callbacks,
-                streaming=True,
-            ),
-            question_generator=self.knowledge_qa.get_question_generation_llm(),
-            verbose=False,
-            rephrase_question=False,
-            return_source_documents=True,
+        conversational_qa_chain = self.knowledge_qa.get_chain()
+        transformed_history, streamed_chat_history = (
+            self.initialize_streamed_chat_history(chat_id, question)
         )
-
-        transformed_history = format_chat_history(history)
-
         response_tokens = []
+        sources = []
 
-        async def wrap_done(fn: Awaitable, event: asyncio.Event):
-            try:
-                return await fn
-            except Exception as e:
-                logger.error(f"Caught exception: {e}")
-                return None  # Or some sentinel value that indicates failure
-            finally:
-                event.set()
-
-        prompt_content = self.prompt_to_use.content if self.prompt_to_use else None
-        run = asyncio.create_task(
-            wrap_done(
-                qa.acall(
-                    {
-                        "question": question.question,
-                        "chat_history": transformed_history,
-                        "custom_personality": prompt_content,
-                    }
+        async for chunk in conversational_qa_chain.astream(
+            {
+                "question": question.question,
+                "chat_history": transformed_history,
+                "custom_personality": (
+                    self.prompt_to_use.content if self.prompt_to_use else None
                 ),
-                callback.done,
-            )
-        )
+            }
+        ):
+            if chunk.get("answer"):
+                logger.info(f"Chunk: {chunk}")
+                response_tokens.append(chunk["answer"].content)
+                streamed_chat_history.assistant = chunk["answer"].content
+                yield f"data: {json.dumps(streamed_chat_history.dict())}"
+            if chunk.get("docs"):
+                sources = chunk["docs"]
 
+        sources_list = generate_source(sources, self.brain_id)
+        if not streamed_chat_history.metadata:
+            streamed_chat_history.metadata = {}
+            # Serialize the sources list
+        serialized_sources_list = [source.dict() for source in sources_list]
+        streamed_chat_history.metadata["sources"] = serialized_sources_list
+        yield f"data: {json.dumps(streamed_chat_history.dict())}"
+        self.save_answer(question, response_tokens, streamed_chat_history, save_answer)
+
+    def initialize_streamed_chat_history(self, chat_id, question):
+        history = chat_service.get_chat_history(self.chat_id)
+        transformed_history = format_chat_history(history)
         brain = brain_service.get_brain_by_id(self.brain_id)
 
-        if save_answer:
-            streamed_chat_history = chat_service.update_chat_history(
-                CreateChatHistory(
-                    **{
-                        "chat_id": chat_id,
-                        "user_message": question.question,
-                        "assistant": "",
-                        "brain_id": brain.brain_id,
-                        "prompt_id": self.prompt_to_use_id,
-                    }
-                )
-            )
-
-            streamed_chat_history = GetChatHistoryOutput(
+        streamed_chat_history = chat_service.update_chat_history(
+            CreateChatHistory(
                 **{
-                    "chat_id": str(chat_id),
-                    "message_id": streamed_chat_history.message_id,
-                    "message_time": streamed_chat_history.message_time,
+                    "chat_id": chat_id,
                     "user_message": question.question,
                     "assistant": "",
-                    "prompt_title": (
-                        self.prompt_to_use.title if self.prompt_to_use else None
-                    ),
-                    "brain_name": brain.name if brain else None,
-                    "brain_id": str(brain.brain_id) if brain else None,
-                    "metadata": self.metadata,
+                    "brain_id": brain.brain_id,
+                    "prompt_id": self.prompt_to_use_id,
                 }
             )
-        else:
-            streamed_chat_history = GetChatHistoryOutput(
-                **{
-                    "chat_id": str(chat_id),
-                    "message_id": None,
-                    "message_time": None,
-                    "user_message": question.question,
-                    "assistant": "",
-                    "prompt_title": (
-                        self.prompt_to_use.title if self.prompt_to_use else None
-                    ),
-                    "brain_name": brain.name if brain else None,
-                    "brain_id": str(brain.brain_id) if brain else None,
-                    "metadata": self.metadata,
-                }
-            )
+        )
 
-        try:
-            async for token in callback.aiter():
-                logger.debug("Token: %s", token)
-                response_tokens.append(token)
-                streamed_chat_history.assistant = token
-                yield f"data: {json.dumps(streamed_chat_history.dict())}"
-        except Exception as e:
-            logger.error("Error during streaming tokens: %s", e)
-        try:
-            # Python
+        streamed_chat_history = GetChatHistoryOutput(
+            **{
+                "chat_id": str(chat_id),
+                "message_id": streamed_chat_history.message_id,
+                "message_time": streamed_chat_history.message_time,
+                "user_message": question.question,
+                "assistant": "",
+                "prompt_title": (
+                    self.prompt_to_use.title if self.prompt_to_use else None
+                ),
+                "brain_name": brain.name if brain else None,
+                "brain_id": str(brain.brain_id) if brain else None,
+                "metadata": self.metadata,
+            }
+        )
 
-            # Await the run
-            result = await run
+        return transformed_history, streamed_chat_history
 
-            sources_list = generate_source(result, brain)
-            # Create metadata if it doesn't exist
-            if not streamed_chat_history.metadata:
-                streamed_chat_history.metadata = {}
-                # Serialize the sources list
-            serialized_sources_list = [source.dict() for source in sources_list]
-            streamed_chat_history.metadata["sources"] = serialized_sources_list
-            yield f"data: {json.dumps(streamed_chat_history.dict())}"
-
-        except Exception as e:
-            logger.error("Error processing source documents: %s", e)
-
-        # Combine all response tokens to form the final assistant message
+    def save_answer(
+        self, question, response_tokens, streamed_chat_history, save_answer
+    ):
         assistant = "".join(response_tokens)
 
         try: